Month Picker Finished
At this point, you can select a month with the new month picker, it shows the 12 months of the year in a smaller way. It uses grid for making the layout of the month container and it also has an awesome animation for displaying the month picker. The only thing thaht I think i could change is the colors used for this month picker. The rest is perfect <3