Upload Section looks good and works fine
At this point the upload section works very good <3. When you press the upload file button and you upload a file it automatically submits the file. The only problem that has to be solved is don't go to another page when the upload has finished. This is a Checkpoint for making a feature branch in order to show in the same page all the rows from the person whose tax recepit has been uploaded