Table can now be displayed
If you upload a file, it shows you a table where you can see all the months corresponding to the person's tax receipt it has been uploaded. But I have to change the way data is displayed. Maybe focusing more on the months rather to the person's name and rfc