Drag N Drop Added & Works Perfect
Drag N Drop has been added and it also validates if the file uploaded is a PDF file. The Month Picker has been removed, the month can be retrieved with the same Python Script that was used for decoding the QR code. Alert needs to be created in order to notify user if the File is not a PDF file and also for notifying if the Tax Receipt added has already been uploaded