Loading Page Animation Already Works for Firefox
I have added the snap.svg library for animating the SVG. This makes the animation possible in Firefox because before using the library I was using the 'd' attribute in CSS and that only works for Chrome. So, now the snap.svg library takes care of the animation. And also, I've tried to use the gsap library for animating the Loading Square but it was more complicated than using CSS keyframes, so I left the code I tried to use with gsap in the 'loadingAnimation.js' file (it is commented)